backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.32.3
numpy>=1.26.0
nltk==3.9.1
phonikud>=0.4.0
flask-limiter==3.8.0
python-dotenv==1.0.1
gunicorn==21.2.0
